requirements.txt
Django==4.1
boto3==1.26.148
django-extensions==3.2.1
psycopg2-binary==2.9.6
requests==2.31.0
whitenoise==6.4.0
mysql-connector-python==8.0.33
djangorestframework==3.12.4
pytz==2023.3
pynamodb
kafka-python
django-admin-extra-buttons
ruff
